Convert abci::ValidatorUpdate to validator::Update
